requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
langchain-text-splitters==0.3.0
jiwer==3.0.4
PyMuPDF==1.24.10
Pillow==10.4.0
numpy==1.26.4
